+++ src/runner.rs
@@ -1,16 +1,10 @@
 use anyhow::Result;
-use directories::ProjectDirs;
 use futures::lock::Mutex;
-use std::path::Path;
 use std::{collections::VecDeque, sync::Arc, time::Duration};
 use tokio::time::sleep;
 
 use crate::app::{Event, ORG};
-
-lazy_static::lazy_static! {
-    static ref DIRS: ProjectDirs = ProjectDirs::from("org", "scverse", "scverse benchmark").expect("No Home dir");
-    static ref CACHE_DIR: &'static Path = DIRS.cache_dir();
-}
+use crate::git::sync_repo;
 
 pub(crate) async fn runner(events: Arc<Mutex<VecDeque<Event>>>) -> Result<()> {
     loop {
@@ -35,20 +29,4 @@ async fn run(repo: String, branch: String) -> Result<()> {
     tracing::info!("Synced repo to {:?}", repo.path());
     // TODO: run
     Ok(())
-}
-
-fn sync_repo(repo: &str, branch: &str) -> Result<git2::Repository> {
-    let path = CACHE_DIR.join(repo);
-    let repo = if path.is_dir() {
-        let r = git2::Repository::open(path)?;
-        // TODO: fetch
-        // r.set_head(branch)?; // TODO: use correct ref
-        r.checkout_head(None)?;
-        r
-    } else {
-        git2::build::RepoBuilder::new()
-            .branch(branch)
-            .clone(&format!("https://github.com/{ORG}/{repo}"), &path)?
-    };
-    Ok(repo)
 }
